% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.2026.01.22.R
\name{tax.shared.plot}
\alias{tax.shared.plot}
\title{Visualize Shared and Unique Taxa Between Groups}
\usage{
tax.shared.plot(
  tax.shared.table_result,
  log = T,
  min.obs = 1,
  color1 = "orange",
  color2 = "purple",
  family.name.length = 2,
  round.to = 3
)
}
\arguments{
\item{log}{Logical; if TRUE, apply log10(x + 1) transformation to the matrix and plot axis. Default is TRUE.}

\item{tax.res}{result from tax.shared.table}
}
\description{
This function visualizes differences in taxonomic abundance between two groups
by creating a horizontal barplot of shared and unique taxa, based on the results from
\code{tax.shared.table()}. It supports log transformation and subsetting to specific taxa.
}
\details{
This function builds upon \code{tax.shared.table()} to generate a visual summary of taxa shared and unique to two groups. Bar lengths represent median abundance (or presence) for each taxon in the respective group.
}
